requirements.txt
streamlit==1.38.0
requests==2.32.3
pandas==2.2.2
matplotlib==3.9.2
plotly==5.23.0
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pandas==2.2.2
PyMuPDF==1.24.9
matplotlib==3.9.2
seaborn==0.13.2
openpyxl==3.1.5
scikit-learn==1.5.1
faiss-cpu==1.8.0.post1
sentence-transformers==3.1.1
